requirements.txt
openai 
streamlit
PyPDF2
langchain
python-dotenv

-e .